+++ Java/Android/OpenGLES/source/core/src/loon/action/sprite/SpriteBatch.java
@@ -1747,7 +1747,7 @@ public class SpriteBatch implements LRelease {
 		if (spriteFont != null) {
 			setFont(spriteFont);
 		}
-		int heigh = (int) ((spriteFont.getHeight() - 2));
+		int heigh = ((spriteFont.getHeight() - 2));
 		if (rotation == 0f) {
 			drawString(text, px - (originx * scale), (py + heigh)
 					- (originy * scale), scale, scale, originx, originy,
@@ -1765,7 +1765,7 @@ public class SpriteBatch implements LRelease {
 		if (spriteFont != null) {
 			setFont(spriteFont);
 		}
-		int heigh = (int) ((spriteFont.getHeight() - 2));
+		int heigh = ((spriteFont.getHeight() - 2));
 		if (rotation == 0f) {
 			drawString(text, position.x - (origin.x * scale),
 					(position.y + heigh) - (origin.y * scale), scale, scale,
@@ -1783,7 +1783,7 @@ public class SpriteBatch implements LRelease {
 		if (spriteFont != null) {
 			setFont(spriteFont);
 		}
-		int heigh = (int) (spriteFont.getHeight() - 2);
+		int heigh = (spriteFont.getHeight() - 2);
 		drawString(text, position.x, (position.y + heigh), 1f, 1f, 0f, 0f, 0f,
 				color);
 		setFont(old);
@@ -1795,7 +1795,7 @@ public class SpriteBatch implements LRelease {
 		if (spriteFont != null) {
 			setFont(spriteFont);
 		}
-		int heigh = (int) (spriteFont.getHeight() - 2);
+		int heigh = (spriteFont.getHeight() - 2);
 		drawString(text, x, (y + heigh), 1f, 1f, 0f, 0f, 0f, color);
 		setFont(old);
 	}
@@ -1806,7 +1806,7 @@ public class SpriteBatch implements LRelease {
 		if (spriteFont != null) {
 			setFont(spriteFont);
 		}
-		int heigh = (int) ((spriteFont.getHeight() - 2));
+		int heigh = ((spriteFont.getHeight() - 2));
 		if (rotation == 0f) {
 			drawString(text, position.x - (origin.x * scale.x),
 					(position.y + heigh) - (origin.y * scale.y), scale.x,
