+++ Java/Android/OpenGLES/source/core/src/loon/core/graphics/component/Speed.java
@@ -84,17 +84,17 @@ public class Speed {
 	public void set(float direction, float length) {
 		this.length = length;
 		this.direction = direction;
-		this.dx = (float) (length * MathUtils.cos(MathUtils
+		this.dx = (length * MathUtils.cos(MathUtils
 				.toRadians(direction)));
-		this.dy = (float) (length * MathUtils.sin(MathUtils
+		this.dy = (length * MathUtils.sin(MathUtils
 				.toRadians(direction)));
 	}
 
 	public void setDirection(float direction) {
 		this.direction = direction;
-		this.dx = (float) (this.length * MathUtils.cos(MathUtils
+		this.dx = (this.length * MathUtils.cos(MathUtils
 				.toRadians(direction)));
-		this.dy = (float) (this.length * MathUtils.sin(MathUtils
+		this.dy = (this.length * MathUtils.sin(MathUtils
 				.toRadians(direction)));
 	}
 
@@ -103,7 +103,7 @@ public class Speed {
 		this.dy += other.dy;
 		this.direction = (int) MathUtils.toDegrees(MathUtils.atan2(this.dy,
 				this.dx));
-		this.length = (float) MathUtils.sqrt(this.dx * this.dx + this.dy
+		this.length = MathUtils.sqrt(this.dx * this.dx + this.dy
 				* this.dy);
 	}
